+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - increased width of the tooltip in MPC tool description for better readability
 - added support for opening updated files diff in a new window
 - filter extension providers when saving them in store
+- correctly loading OpenAI-compatible models when API key is left empty
 
 ## [0.60.0]